# docs/swagger.yaml
basePath: /api/v1
consumes:
- application/json
definitions:
...
  choices.PaymentPurpose:
    enum:
    - COINS
    - SUBSCRIPTION
    type: string
    x-enum-varnames:
    - PP_COINS
    - PP_SUB
  choices.PaymentStatus:
    enum:
    - PENDING
    - SUCCEEDED
    - FAILED
    - CANCELED
    type: string
    x-enum-varnames:
    - PSPENDING
    - PSSUCCEEDED
    - PSFAILED
    - PSCANCELED
  choices.PaymentType:
    enum:
    - GOOGLE PAY
    - STRIPE
    - PAYPAL
    type: string
    x-enum-varnames:
    - PTYPE_GPAY
    - PTYPE_STRIPE
    - PTYPE_PAYPAL
...
  choices.SubscriptionTypeChoice:
    enum:
    - MONTHLY
    - ANNUAL
    type: string
    x-enum-varnames:
    - ST_MONTHLY
    - ST_ANNUAL
...
  schemas.BuyCoinSchema:
    properties:
      coin_id:
        example: 19e8bd22-fab1-4bb4-ba82-77c41bea6b99
        type: string
      quantity:
        description: PaymentType choices.PaymentType `json:"payment_type" validate:"required,payment_type_validator"
          example:"STRIPE"` // This should be stripe by default
        example: 2
        type: integer
    required:
    - coin_id
    - quantity
    type: object
...
  schemas.CoinSchema:
    properties:
      amount:
        example: 5
        type: integer
      id:
        example: 19e8bd22-fab1-4bb4-ba82-77c41bea6b99
        type: string
      price:
        example: 10.45
        type: number
    type: object
  schemas.CoinsResponseSchema:
    properties:
      data:
        items:
          $ref: '#/definitions/schemas.CoinSchema'
        type: array
      message:
        example: Data fetched/created/updated/deleted
        type: string
      status:
        example: success
        type: string
    type: object
...
  schemas.CreateSubscriptionSchema:
    properties:
      payment_method_token:
        type: string
      subtype:
        $ref: '#/definitions/choices.SubscriptionTypeChoice'
    required:
    - payment_method_token
    - subtype
    type: object
...
  schemas.PaymentResponseSchema:
    properties:
      data:
        $ref: '#/definitions/schemas.TransactionSchema'
      message:
        example: Data fetched/created/updated/deleted
        type: string
      status:
        example: success
        type: string
    type: object
...
  schemas.SubscriptionPlanResponseSchema:
    properties:
      data:
        $ref: '#/definitions/schemas.SubscriptionPlanSchema'
      message:
        example: Data fetched/created/updated/deleted
        type: string
      status:
        example: success
        type: string
    type: object
  schemas.SubscriptionPlanSchema:
    properties:
      amount:
        type: number
      subtype:
        $ref: '#/definitions/choices.SubscriptionTypeChoice'
    required:
    - amount
    - subtype
    type: object
...
  schemas.SubscriptionPlansResponseSchema:
    properties:
      data:
        items:
          $ref: '#/definitions/schemas.SubscriptionPlanSchema'
        type: array
      message:
        example: Data fetched/created/updated/deleted
        type: string
      status:
        example: success
        type: string
    type: object
...
  schemas.TransactionSchema:
    properties:
      amount:
        example: 10.35
        type: number
      amount_total:
        example: 30.35
        type: number
      client_secret:
        type: string
      coins:
        example: 10
        type: integer
      coins_total:
        example: 30
        type: integer
      created_at:
        type: string
      payment_purpose:
        allOf:
        - $ref: '#/definitions/choices.PaymentPurpose'
        example: SUBSCRIPTION
      payment_status:
        $ref: '#/definitions/choices.PaymentStatus'
      payment_type:
        allOf:
        - $ref: '#/definitions/choices.PaymentType'
        example: STRIPE
      quantity:
        example: 10
        type: integer
      reference:
        type: string
      updated_at:
        type: string
    type: object
  schemas.TransactionsResponseDataSchema:
    properties:
      current_page:
        example: 1
        type: integer
      last_page:
        example: 100
        type: integer
      per_page:
        example: 100
        type: integer
      transactions:
        items:
          $ref: '#/definitions/schemas.TransactionSchema'
        type: array
    type: object
  schemas.TransactionsResponseSchema:
    properties:
      data:
        $ref: '#/definitions/schemas.TransactionsResponseDataSchema'
      message:
        example: Data fetched/created/updated/deleted
        type: string
      status:
        example: success
        type: string
    type: object
...
  utils.ErrorResponse:
    properties:
      code:
        type: string
      data:
        additionalProperties:
          type: string
        type: object
      message:
        type: string
      status:
        type: string
    type: object
info:
  contact: {}
  description: |-
    ## LitPAD API built with Fiber and GORM

    ### WEBSOCKETS:

    #### Notifications

    - URL: `wss://{host}/api/v1/ws/notifications`

    - Requires authorization, so pass in the Bearer Authorization header.

    - You can only read and not send notification messages into this socket.
  title: LITPAD API
  version: "1.0"
paths:
...
  /wallet/coins:
    get:
      description: This endpoint returns all available coins for sale
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.CoinsResponseSchema'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: View Available Coins
      tags:
      - Wallet
    post:
      description: This endpoint allows a user to buy coins
      parameters:
      - description: Payment object
        in: body
        name: coin
        required: true
        schema:
          $ref: '#/definitions/schemas.BuyCoinSchema'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.PaymentResponseSchema'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Buy Coins
      tags:
      - Wallet
...
  /wallet/plans:
    get:
      consumes:
      - application/json
      description: Retrieves a list of available subscription plans.
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved list of plans
          schema:
            $ref: '#/definitions/schemas.SubscriptionPlansResponseSchema'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: List Available Subscription Plans
      tags:
      - Wallet
    put:
      description: This endpoint allows an admin to change the amount of a plan
      parameters:
      - description: Plan data
        in: body
        name: plan
        required: true
        schema:
          $ref: '#/definitions/schemas.SubscriptionPlanSchema'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.SubscriptionPlanResponseSchema'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update A Plan Amount
      tags:
      - Wallet
  /wallet/subscription:
    post:
      description: This endpoint allows a user to create a subscription for books
      parameters:
      - description: Payment object
        in: body
        name: subscription
        required: true
        schema:
          $ref: '#/definitions/schemas.CreateSubscriptionSchema'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.PaymentResponseSchema'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Subscribe
      tags:
      - Wallet
  /wallet/transactions:
    get:
      description: This endpoint returns all transactions of a user
      parameters:
      - default: 1
        description: Current Page
        in: query
        name: page
        type: integer
      - description: Payment Status
        in: query
        name: payment_status
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.TransactionsResponseSchema'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: View Current Transactions
      tags:
      - Wallet
produces:
- application/json
security:
- BearerAuth: []
securityDefinitions:
  BearerAuth:
    description: Type 'Bearer jwt_string' to correctly set the API Key
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"